requirements.txt
chardet==2.2.1
colorama==0.2.5
html5lib==0.999
pycurl==7.43.0.6
PyMySQL==1.0.2
python-decouple==3.4
requests==2.2.1
six==1.5.2
urllib3==1.7.1